drm/sun4i: Set the coef_rdy bit right after the coef have been set
The COEF_RDY bit is used to tell the hardware that new FIR filters
coefficients have been written to the registers and that the hardware
should take them into account starting next frame.

The current code doesn't set that bit right after, but before, which
obviously doesn't work. Move it where it belongs.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>