ARM: dts: sun4i-a10: Add Video Engine and reserved memory nodes
This ass nodes for the Video Engine and the associated reserved memory
for the A10. Up to 96 MiB of memory are dedicated to the CMA pool.

The VPU can only map the first 256 MiB of DRAM, so the reserved memory
pool has to be located in that area. Following Allwinner's decision in
downstrean software, the last 96 MiB of the first 256 MiB of RAM are
reserved for this purpose.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>